Renamed checkoutdotcom to checkoutcom
