+++ README.md
@@ -46,7 +46,7 @@ $ export CYBERSOURCE_ACCOUNT="YOUR_CYBS_ACCOUNT"
 $ export CYBERSOURCE_API_KEY="YOUR_CYBS_KEY"
 $ export CYBERSOURCE_SHARED_SECRET="YOUR_CYBS_SECRET"
 $ export NMI_SECURITY_KEY="YOUR_NMI_PRIVATE_KEY"
-$ export CHECKOUTDOTCOM_TEST_KEY="YOUR_CHECKOUTDOTCOM_PRIVATE_KEY"
+$ export CHECKOUTCOM_TEST_KEY="YOUR_CHECKOUTCOM_PRIVATE_KEY"
 ```
 
 Then run tests with: `go test ./integration-tests/`
